refactor: calculator 도메인을 분리
- 정수 숫자의 합을 단위테스트하기 용이해짐
- calculator 도메인에 대한 독립적인 단위 테스트를 할 수 있음
- 예외 케이스인 maxInt + maxInt에 대해 처리

+++ src/main/java/calculator/application/ExtractService.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface ExtractService {
 
-    List<Integer> extractNumbers(List<String> data);
+    List<Long> extractNumbers(List<String> data);
 
 }
