+++ src/main/java/calculator/ui/view/CalculatorOutputView.java
@@ -4,7 +4,7 @@ public class CalculatorOutputView {
 
     private static final StringBuilder BUFFER = new StringBuilder();
 
-    public void splitAndSumOutput(int result) {
+    public void splitAndSumOutput(long result) {
         BUFFER.append("결과 : ").append(result).append('\n');
         System.out.println(BUFFER);
         bufferClear();
